requirements.txt
Django==1.10.6
argparse==1.2.1
djangorestframework==3.6.2
wsgiref==0.1.2
